requirements.txt
aiosqlite>=0.19.0
dnspython>=2.4.0
pandas>=2.0.0
polars>=0.20.0
python-dotenv>=1.0.0
requests>=2.28.0
